requirements.txt
python-dotenv
apscheduler
fastapi
uvicorn[gunicorn]
gunicorn